ai_services_apis/urls.py
"""
URL configuration for ai_services_apis project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path
from django.conf.urls import url
from rest_framework.authtoken.views import obtain_auth_token
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView

from ai_services.file_2_text.file_to_text_view_v1 import File2TextConverterAPI
from ai_services.views import HealthCheck

urlpatterns = [
    path('admin/', admin.site.urls),
    path('/', admin.site.urls),
    url(r'^health_check/', HealthCheck, name='health_check'),
    url(r'^login/', obtain_auth_token, name='login'),
    path('api/v1/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/v1/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/v1/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    url(r'^api/v1/file2text/$', File2TextConverterAPI, name="file2text"),

]