requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary==2.9.9
boto3==1.17.62
gunicorn==20.1.0
environs==9.3.2
djangorestframework-simplejwt==4.8.0
django-filter==21.1
pandas==2.2.2
tika==1.24
spacy==3.7.5
pytesseract==0.3.13
Pillow==10.4.0
pdfplumber==0.11.6
pyspellchecker==0.8.1
python-magic==0.4.27
filetype==1.2.0
docx2txt==0.9
